cambios en la ruta product (funcionando)
